complete the notifications view
