+++ app/Http/Controllers/NotificationsController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
+use App\Notification;
+
 use App\Http\Requests;
 
 class NotificationsController extends Controller
@@ -15,19 +17,21 @@ class NotificationsController extends Controller
     public function index()
     {
         $notifications = Notification::all();
+
         return view('notifications.index', compact('notifications'));
     }
 
-    public function show($id)
-    {
-        try {
-            $notification = Notification::findOrfail($id);
-        } catch(ModelNotFoundException $e) {
-            return view('errors.404');
-        }
-
-        return view('notifications.show', compact('notification'));
-    }
+//    public function show($id)
+//    {
+//
+//        try {
+//            $notification = Notification::findOrfail($id);
+//        } catch(ModelNotFoundException $e) {
+//            return view('errors.404');
+//        }
+//
+//        return view('notifications.show', compact('notification'));
+//    }
 
 
 }
