+++ app/Notification.php
@@ -6,5 +6,18 @@ use Illuminate\Database\Eloquent\Model;
 
 class Notification extends Model
 {
-    //
+
+    protected $fillable = [
+        'read', 'messsage', 'user_id'
+    ];
+
+    protected $table = 'notifications';
+
+
+    public function user()
+    {
+        return $this->belongsTo('App\User');
+    }
+
+
 }
