+++ resources/views/notifications/index.blade.php
@@ -1,0 +1,45 @@
+@extends('admin_template')
+
+@section('content')
+
+    <ul class="menu">
+        <li id="empty_li_main" hidden></li>
+        @foreach($notifications as $notification)
+
+            <li>
+                <a href="#">
+                    {{ $notification->message }}
+                    at {{ $notification->created_at }}
+                </a>
+            </li>
+
+
+        @endforeach
+    </ul>
+
+
+
+    <script src="https://cdn.socket.io/socket.io-1.3.5.js"></script>
+    <script>
+        //var socket = io('http://localhost:3000');
+        var socket = io('http://localhost:3000');
+        socket.on("notifications-channel:App\\Events\\NotificationsEvent", function(message){
+            // increase the power everytime we load test route
+            // $('#notification').text($('#notification').text() + "\n" + message.data.message);
+            console.log("hello2");
+            $('#empty_li_main').after("<li><a href='#'> <i class='fa fa-users text-aqua'></i>" +  message.data.message + " at " + message.data.at + " </a> </li>");
+        });
+    </script>
+
+
+
+
+
+@endsection
+
+
+
+
+
+
+
